move some from fluid.io to static.io
