+++ test/ir/inference/quant_dequant_test.py
@@ -25,7 +25,7 @@ from paddle import fluid
 from paddle.fluid import Program, Variable, core
 from paddle.fluid.core import AnalysisConfig, create_paddle_predictor
 from paddle.fluid.framework import IrGraph
-from paddle.fluid.io import append_fetch_ops, prepend_feed_ops
+from paddle.static.io import append_fetch_ops, prepend_feed_ops
 from paddle.static.quantization import (
     AddQuantDequantPass,
     OutScaleForInferencePass,
